+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `jiff`
   - human timespan strings for `TimeSpan` and `SignedDuration` objects:
     - `ry.TimeSpan.parse("P2M10DT2H30M").string(human=True) == "2mo 10d 2h 30m"`
+    - `ry.SignedDuration.parse("PT2H30M").string(human=True) == "2h 30m"`
 
 ___